requirements.txt
asgiref==3.8.1
certifi==2024.12.14
dj-database-url==2.3.0
Django==5.2.7
django-allauth==65.10.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
